Minor syntax fix
